builder/image_augmentation/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64.0.0", "wheel"]

[project]
name = "image_augmentation"
dynamic = ["version"]
dependencies = [
  "nvidia-nat[redis,mcp,langchain,phoenix,async_endpoints]==1.6.0-dev",
  "httpx>=0.28",
  "redis~=4.3.6"
]
requires-python = ">=3.11,<3.14"
description = "Image augmentation function for NeMo Agent Toolkit"
classifiers = ["Programming Language :: Python"]

[project.entry-points.'nat.components']
image_augmentation = "image_augmentation.register"

[tool.setuptools]
packages = ["image_augmentation"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
image_augmentation = ["configs/*.yml", "configs/*.yaml"]

[tool.uv]
config-settings = { editable_mode = "compat" }
